refactor(mcp): unify MCP onto one configurable toolsets surface
Per PR design review (bplatz): expose MCP as a single configurable server
with selectable toolsets, not one server per feature — so the shape holds
when database-over-stdio arrives.

New `fluree-db-mcp` crate hosts a unified `FlureeMcpService`: each toolset
(`memory`, `docs`) is a named rmcp `#[tool_router]` block on one service
type, and `new()` composes only the enabled toolsets' routers into the
router that `#[tool_handler]` dispatches against (so tools/list reflects
exactly the selection). The per-crate rmcp wrappers (`MemoryToolService`,
`DocsToolService`) are removed; the domain crates keep only their logic and
drop their `mcp` features.

Command surface:
- `fluree mcp init [--ide] [--toolsets memory,docs|all]` — writes one IDE
  entry named `fluree`, migrating away legacy `fluree-memory`/`fluree-docs`
  entries.
- `fluree mcp serve [--toolsets …]` — bare serve defaults to memory for
  back-compat; docs-only is stateless and needs no `.fluree/` directory.
- `fluree mcp status` — reports installed toolsets per detected IDE.
- `fluree docs serve` removed; `fluree docs search|get|examples|tree` kept.
- `fluree mcp install`, `fluree memory mcp-install`, `fluree memory init`
  kept as hidden back-compat aliases (init no longer creates the store; it
  is lazy-created on first use).

Tested: unified-service composition (per-toolset tool lists), installer
single-entry + legacy migration, and end-to-end stdio (initialize +
tools/list + docs_search tools/call) for docs/memory/all.
