+++ Cargo.toml
@@ -38,6 +38,7 @@ members = [
     "fluree-db-cli",
     "fluree-db-memory",
     "fluree-db-docs",
+    "fluree-db-mcp",
     "fluree-bench-support",
     "fluree-db-consensus",
 ]
