+++ fluree-db-cli/Cargo.toml
@@ -45,14 +45,14 @@ fluree-vocab = { path = "../fluree-vocab" }
 fluree-db-query = { path = "../fluree-db-query" }
 fluree-db-binary-index = { path = "../fluree-db-binary-index" }
 
-# Developer memory layer
-fluree-db-memory = { path = "../fluree-db-memory", features = ["mcp"] }
+# Developer memory layer (store + CLI subcommands)
+fluree-db-memory = { path = "../fluree-db-memory" }
 
-# Embedded docs lookup (CLI subcommand + standalone fluree-docs MCP server)
-fluree-db-docs = { path = "../fluree-db-docs", features = ["mcp"] }
+# Embedded docs lookup (human `fluree docs` CLI subcommand)
+fluree-db-docs = { path = "../fluree-db-docs" }
 
-# MCP stdio server
-rmcp = { workspace = true }
+# Unified MCP service (stdio) — selectable memory/docs toolsets
+fluree-db-mcp = { path = "../fluree-db-mcp" }
 
 # Server embedding (optional — enables `fluree server` subcommands)
 fluree-db-server = { path = "../fluree-db-server", optional = true }
