+++ fluree-db-cli/src/lib.rs
@@ -625,57 +625,67 @@ pub async fn run(cli: Cli) -> error::CliResult<()> {
         }
 
         Commands::Mcp { action } => match action {
-            cli::McpAction::Serve { transport } => {
-                // IDEs may spawn `fluree mcp serve` from a cwd that is not inside
-                // a project with a local `.fluree/` directory. In that case, fall
-                // back to global directories (creating them if needed) so the MCP
-                // server can still start and expose tools.
+            cli::McpAction::Serve {
+                transport,
+                toolsets,
+            } => {
+                let toolsets = commands::mcp::parse_toolsets(&toolsets)?;
+
+                // Only resolve (and, if needed, create) a Fluree directory when
+                // the memory toolset is enabled — a docs-only server is stateless
+                // and must not require or create a `.fluree/` directory.
                 //
-                // IMPORTANT: This must not print to stdout/stderr (stdio transport
-                // uses stdout for JSON-RPC).
-                let fluree_dir = if let Some(p) = config_path {
-                    config::require_fluree_dir(Some(p))?
-                } else if let Some(local) = config::find_fluree_dir() {
-                    local
+                // IMPORTANT: this must not print to stdout/stderr (stdio
+                // transport uses stdout for JSON-RPC).
+                let fluree_dir = if toolsets.contains(&fluree_db_mcp::Toolset::Memory) {
+                    Some(resolve_mcp_fluree_dir(config_path)?)
                 } else {
-                    let global = FlureeDir::global().ok_or_else(|| {
-                        error::CliError::Config("cannot determine global directories".into())
-                    })?;
-
-                    // For split global dirs, use an absolute data-dir storage path
-                    // so the server finds the right directory regardless of cwd.
-                    let storage_override = if !global.is_unified() {
-                        let path = global.data_dir().join("storage");
-                        let path_str = path.to_str().ok_or_else(|| {
-                            error::CliError::Config(format!(
-                                "data directory path is not valid UTF-8: {}",
-                                path.display()
-                            ))
-                        })?;
-                        Some(path_str.to_owned())
-                    } else {
-                        None
-                    };
-
-                    let template = generate_config_template_for(
-                        ConfigFormat::Toml,
-                        storage_override.as_deref(),
-                    );
-
-                    // Create minimal directory structure if missing.
-                    config::init_fluree_dir(&global, &template, ConfigFormat::Toml.filename())?;
-                    global
+                    None
                 };
-                commands::mcp_serve::run(&transport, &fluree_dir).await
+                commands::mcp::serve::run(&transport, &toolsets, fluree_dir.as_ref()).await
             }
-            // Install needs no project directory — the docs server is stateless
-            // and the memory server creates its own store on first use.
-            cli::McpAction::Install { ide, server } => {
-                commands::memory::mcp_install(ide.as_deref(), &server)
+            // init/install need no project directory — the server lazy-inits its
+            // store and the docs toolset is stateless.
+            cli::McpAction::Init { ide, toolsets } | cli::McpAction::Install { ide, toolsets } => {
+                commands::mcp::init(ide.as_deref(), &toolsets)
             }
+            cli::McpAction::Status => commands::mcp::status(),
         },
 
         // Docs are embedded in the binary — no project directory needed.
-        Commands::Docs { action } => commands::docs::run(action).await,
+        Commands::Docs { action } => commands::docs::run(action),
     }
+}
+
+/// Resolve the Fluree directory for an MCP server that includes the memory
+/// toolset. IDEs may spawn `fluree mcp serve` from a cwd outside a project with
+/// a local `.fluree/`, so fall back to (and create) global directories.
+fn resolve_mcp_fluree_dir(config_path: Option<&std::path::Path>) -> error::CliResult<FlureeDir> {
+    if let Some(p) = config_path {
+        return config::require_fluree_dir(Some(p));
+    }
+    if let Some(local) = config::find_fluree_dir() {
+        return Ok(local);
+    }
+    let global = FlureeDir::global()
+        .ok_or_else(|| error::CliError::Config("cannot determine global directories".into()))?;
+
+    // For split global dirs, use an absolute data-dir storage path so the server
+    // finds the right directory regardless of cwd.
+    let storage_override = if !global.is_unified() {
+        let path = global.data_dir().join("storage");
+        let path_str = path.to_str().ok_or_else(|| {
+            error::CliError::Config(format!(
+                "data directory path is not valid UTF-8: {}",
+                path.display()
+            ))
+        })?;
+        Some(path_str.to_owned())
+    } else {
+        None
+    };
+
+    let template = generate_config_template_for(ConfigFormat::Toml, storage_override.as_deref());
+    config::init_fluree_dir(&global, &template, ConfigFormat::Toml.filename())?;
+    Ok(global)
 }
