+++ fluree-db-memory/Cargo.toml
@@ -9,7 +9,6 @@ repository.workspace = true
 [features]
 default = ["native"]
 native = ["fluree-db-api/native"]
-mcp = ["dep:rmcp", "dep:schemars"]
 
 [dependencies]
 fluree-db-api = { path = "../fluree-db-api", default-features = false }
@@ -28,10 +27,6 @@ thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 ulid.workspace = true
-
-# MCP support (optional)
-rmcp = { workspace = true, optional = true }
-schemars = { workspace = true, optional = true }
 
 [dev-dependencies]
 tempfile = "3"
